Update dependencies (#498)
* Update dependencies

* Fix type errors

* fix lint

* fix

* fix

+++ src/actions/helpers/running-process.ts
@@ -14,7 +14,7 @@ export default class RunningProcess {
     return this.process != null
   }
   kill() {
-    this.process.kill()
+    this.process && this.process.kill()
   }
 
   reset() {
